컴포넌트 나누기
